User can skip installing R

+++ install.py
@@ -13,7 +13,7 @@ rbaseurl    = "http://cran.r-project.org/src/base/R-3"
 blasurl     = "https://github.com/xianyi/OpenBLAS/zipball/master"
 blasctlurl  = "http://prs.ism.ac.jp/~nakama/SurviveGotoBLAS2/blas_control_on_R/blasctl_0.2.tar.gz"
 
-import os, shutil, urllib2
+import os, shutil, urllib2, sys
 from os import path
 from subprocess import Popen, PIPE
 
@@ -49,6 +49,9 @@ parser.add_argument('-o', '--outdir', default='/usr', help="""In which
                     directory should R be installed (default: %(default)s)""")
 parser.add_argument('--r-opts', default=[], nargs='+', help="""Additional 
                     options to be used when running configure for R""")
+parser.add_argument('-i', '--indir', default=None, help="""What is the path of 
+                    the current R install? This option will then skip installing
+                    R and also skips the use of -o and --r-opts options.""")
 
 args = parser.parse_args()
 
@@ -72,40 +75,43 @@ else:
 ## Install R ##
 ###############
 
-print '\n============'
-print 'Installing R'
+if args.indir is None:
+    args.indir = args.outdir
 
-# Get the url and filename to the latest version of R
-print '\nFetching download url'
-
-br = mechanize.Browser()
-br.open(rbaseurl)
-link = [ l for l in br.links(url_regex="R-") ][-1]
-link = link.absolute_url
-
-print '\nDownloading and Extracting'
-execute("wget -O r_latest.tar.gz %s" % link, "download failure")
-execute("tar -xzvf r_latest.tar.gz", "extraction failure")
-
-print '\nChanging Directories'
-tmp = execute('tar -tzf r_latest.tar.gz | head -n 1', \
-              'could not figure out directory name', \
-              capture_output=True)
-dirname = tmp.strip().strip('/$')
-os.chdir(dirname)
-
-print '\nCompiling and Installing'
-execute('./configure --prefix %s %s' % (args.outdir, " ".join(args.r_opts)), \
-        'configure of R failed')
-execute('make', 'make of R failed')
-execute('make install', 'install of R failed')
-
-print '\nDone, removing source directory'
-os.chdir(startdir)
-shutil.rmtree(dirname)
-os.remove("r_latest.tar.gz")
-
-print '============\n'
+    print '\n============'
+    print 'Installing R'
+    
+    # Get the url and filename to the latest version of R
+    print '\nFetching download url'
+    
+    br = mechanize.Browser()
+    br.open(rbaseurl)
+    link = [ l for l in br.links(url_regex="R-") ][-1]
+    link = link.absolute_url
+    
+    print '\nDownloading and Extracting'
+    execute("wget -O r_latest.tar.gz %s" % link, "download failure")
+    execute("tar -xzvf r_latest.tar.gz", "extraction failure")
+    
+    print '\nChanging Directories'
+    tmp = execute('tar -tzf r_latest.tar.gz | head -n 1', \
+                  'could not figure out directory name', \
+                  capture_output=True)
+    dirname = tmp.strip().strip('/$')
+    os.chdir(dirname)
+    
+    print '\nCompiling and Installing'
+    execute('./configure --prefix %s %s' % (args.outdir, " ".join(args.r_opts)), \
+            'configure of R failed')
+    execute('make', 'make of R failed')
+    execute('make install', 'install of R failed')
+    
+    print '\nDone, removing source directory'
+    os.chdir(startdir)
+    shutil.rmtree(dirname)
+    os.remove("r_latest.tar.gz")
+    
+    print '============\n'
 
 
 ######################
@@ -129,10 +135,13 @@ print '\nCompiling'
 execute('make', 'make of OpenBlas failed')
 
 print '\nInstalling'
-# archive default BLAS library
-rdir1   = path.join(args.outdir, 'lib64', 'R', 'lib')
-rdir2   = path.join(args.outdir, 'lib', 'R', 'lib')
+# archive default BLAS library but first check that required files exist
+rdir1   = path.join(args.indir, 'lib64', 'R', 'lib')
+rdir2   = path.join(args.indir, 'lib', 'R', 'lib')
 rdir    = rdir1 if os.path.isdir(rdir1) else rdir2
+if not path.exists(path.join(rdir, 'libRblas.so')):
+    print "ERROR: Couldn't find the input %s. Make sure R is compiled so blas library is dynamically linked" % path.join(rdir, 'libRblas.so')
+    sys.exit(1)
 shutil.move(path.join(rdir,'libRblas.so'), path.join(rdir,'libRblas.so.keep'))
 # copy over OpenBlas library
 openblas_file = os.path.realpath(os.readlink('libopenblas.so'))
